[CodeCompletion] Split CodeCompletionResult and ContextFreeCodeCompletionResult into their own file
